Implemented means of defining mapping functions
