+++ main.py
@@ -1,6 +1,7 @@
 from api import FlexApp
+from test import mapper
 
-app = FlexApp()
+app = FlexApp(mapper)
 
 if __name__ == '__main__':
    app.run(debug=True)
