+++ test.py
@@ -1,4 +1,11 @@
-import api
+from mapping import JsonMapper
 
-# need to test the JSONMapper class. Try subclassing with the decorator.
-class Test...
+mapper = JsonMapper()
+
+# need to test the JSONMapper class. 
+# Try subclassing with the decorator
+@mapper.maps('/pokemon')
+def pokemon_mapper(self, json):
+    new_json = json
+    new_json['weight'] = 78
+    return new_json
